Singleton keeps high score, deadmenu show new high score

+++ 01_Unity Project/BirdDash/Assets/Scripts/Deliveries/NewTarget.cs
@@ -12,8 +12,8 @@ public class NewTarget : MonoBehaviour
     public Vector3 distance;
     public float threshold;
 
-    public float meal;
-    public float score;
+    public int meal;
+    public int score;
     
 
     GameObject[] allTargets;
